Ken Reorg/Shoryureppa ACMD Sync with new anim
