+++ src/fighter/ken/status.rs
@@ -1,756 +1,15 @@
-use {
-    smash::{
-        lua2cpp::L2CFighterCommon,
-        hash40,
-        phx::Hash40,
-        app::{lua_bind::*, *},
-        lib::{lua_const::*, L2CValue}
-    },
-    smash_script::*,
-    smashline::*,
-    custom_var::*,
-    wubor_utils::{
-        vars::*,
-        table_const::*
-    },
-    super::super::{
-        common::common_status::dash::*,
-        ryu::helper::*
-    }
-};
-
-#[status_script(agent = "ken", status = FIGHTER_RYU_STATUS_KIND_DASH_BACK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_PRE)]
-unsafe fn ken_dashback_pre(fighter: &mut L2CFighterCommon) -> L2CValue {
-    fgc_dashback_pre(fighter)
-}
-
-#[status_script(agent = "ken", status = FIGHTER_RYU_STATUS_KIND_DASH_BACK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_MAIN)]
-unsafe fn ken_dashback_main(fighter: &mut L2CFighterCommon) -> L2CValue {
-    fgc_dashback_main(fighter)
-}
-
-#[status_script(agent = "ken", status = FIGHTER_STATUS_KIND_ATTACK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_MAIN)]
-unsafe fn ken_attack_main(fighter: &mut L2CFighterCommon) -> L2CValue {
-    fighter.sub_status_AttackCommon();
-    if WorkModule::is_flag(fighter.module_accessor, *FIGHTER_RYU_INSTANCE_WORK_ID_FLAG_NEAR_OPPONENT) {
-        WorkModule::set_int64(fighter.module_accessor, 0x10556e6036, *FIGHTER_STATUS_ATTACK_WORK_INT_ATTACK11_MOTION);
-        WorkModule::set_int(fighter.module_accessor, *FIGHTER_LOG_ATTACK_KIND_ATTACK_NEAR, *FIGHTER_RYU_STATUS_ATTACK_INT_LOG_KIND);
-    }
-    else {
-        WorkModule::set_int64(fighter.module_accessor, 0xb4f4e6f8f, *FIGHTER_STATUS_ATTACK_WORK_INT_ATTACK11_MOTION);
-        WorkModule::set_int(fighter.module_accessor, *FIGHTER_LOG_ATTACK_KIND_ATTACK11, *FIGHTER_RYU_STATUS_ATTACK_INT_LOG_KIND);
-    }
-    if !StopModule::is_stop(fighter.module_accessor) {
-        ryu_attack_main_uniq_chk(fighter);
-    }
-    fighter.global_table[SUB_STATUS3].assign(&L2CValue::Ptr(ryu_attack_main_uniq_chk as *const () as _));
-    if !StopModule::is_stop(fighter.module_accessor) {
-        ryu_attack_main_uniq_chk4(fighter, false.into());
-    }
-    fighter.global_table[SUB_STATUS].assign(&L2CValue::Ptr(ryu_attack_main_uniq_chk4 as *const () as _));
-    fighter.sub_shift_status_main(L2CValue::Ptr(ken_attack_main_loop as *const () as _))
-}
-
-unsafe extern "C" fn ken_attack_main_loop(fighter: &mut L2CFighterCommon) -> L2CValue {
-    if !CancelModule::is_enable_cancel(fighter.module_accessor) {
-        if WorkModule::is_flag(fighter.module_accessor, *FIGHTER_RYU_INSTANCE_WORK_ID_FLAG_FINAL_HIT_CANCEL)
-        && AttackModule::is_infliction_status(fighter.module_accessor, *COLLISION_KIND_MASK_SHIELD | *COLLISION_KIND_MASK_HIT)
-        && ryu_final_hit_cancel(fighter, SITUATION_KIND_GROUND.into()).get_bool() {
-            return 1.into();
-        }
-        if WorkModule::is_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_ATTACK_FLAG_HIT_CANCEL)
-        && AttackModule::is_infliction_status(fighter.module_accessor, *COLLISION_KIND_MASK_SHIELD | *COLLISION_KIND_MASK_HIT)
-        && ryu_hit_cancel(fighter, SITUATION_KIND_GROUND.into()).get_bool() {
-            return 1.into();
-        }
-    }
-    if ComboModule::count(fighter.module_accessor) == 1
-    && !CancelModule::is_enable_cancel(fighter.module_accessor) {
-        let current_frame = fighter.global_table[MOTION_FRAME].get_f32();
-        let attack_start_cancel_frame = WorkModule::get_param_float(fighter.module_accessor, hash40("param_private"), hash40("attack_start_cancel_frame"));
-        if current_frame < attack_start_cancel_frame
-        && ryu_kara_cancel(fighter).get_bool() {
-            return 1.into();
-        }
-    }
-    if CancelModule::is_enable_cancel(fighter.module_accessor)
-    && fighter.sub_wait_ground_check_common(false.into()).get_bool() {
-        return 1.into();
-    }
-    let mot = MotionModule::motion_kind(fighter.module_accessor);
-    if [
-        hash40("attack_11_w"),
-        hash40("attack_11_s"),
-        hash40("attack_11_near_s")
-    ].contains(&mot) {
-        if WorkModule::is_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_ATTACK_FLAG_WEAK_CANCEL)
-        && WorkModule::is_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_ATTACK_FLAG_BUTTON_TRIGGER)
-        && ControlModule::check_button_off(fighter.module_accessor, *CONTROL_PAD_BUTTON_ATTACK) {
-            let stick_y = fighter.global_table[STICK_Y].get_f32();
-            let attack_hi3_stick_y = WorkModule::get_param_float(fighter.module_accessor, hash40("common"), hash40("attack_hi3_stick_y"));
-            let cont = if stick_y >= attack_hi3_stick_y {
-                false
-            }
-            else {
-                let attack_lw3_stick_y = WorkModule::get_param_float(fighter.module_accessor, hash40("common"), hash40("attack_lw3_stick_y"));
-                if attack_lw3_stick_y >= stick_y {
-                    false
-                }
-                else {
-                    let stick_x = fighter.global_table[STICK_X].get_f32();
-                    let attack_s3_stick_x = WorkModule::get_param_float(fighter.module_accessor, hash40("common"), hash40("attack_s3_stick_x"));
-                    stick_x < attack_s3_stick_x
-                }
-            };
-            if cont {
-                fighter.change_status(FIGHTER_STATUS_KIND_ATTACK.into(), false.into());
-                return 1.into();
-            }
-        }
-        if WorkModule::is_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_ATTACK_FLAG_SAME_ATTACK_CANCEL) {
-            let button_on_frame = WorkModule::get_int(fighter.module_accessor, *FIGHTER_RYU_STATUS_ATTACK_INT_BUTTON_ON_FRAME);
-            let attack_11_s_button_on_frame = WorkModule::get_param_float(fighter.module_accessor, hash40("param_private"), hash40("attack_11_s_button_on_frame"));
-            if attack_11_s_button_on_frame <= button_on_frame as f32 {
-                fighter.change_status(FIGHTER_STATUS_KIND_ATTACK.into(), false.into());
-                return 1.into();
-            }
-        }
-    }
-    if fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_AIR {
-        fighter.change_status(FIGHTER_STATUS_KIND_FALL.into(), false.into());
-        return 0.into();
-    }
-    if 0 < WorkModule::get_int(fighter.module_accessor, *FIGHTER_STATUS_WORK_ID_INT_RESERVE_ATTACK_MINI_JUMP_ATTACK_FRAME)
-    && !StopModule::is_stop(fighter.module_accessor)
-    && fighter.sub_check_button_jump().get_bool() {
-        // let mot = MotionModule::motion_kind(fighter.module_accessor);
-        // let func = ryu_get_attack_cancel_function(fighter, mot.into());
-        // MotionAnimcmdModule::call_script_single(
-        //     fighter.module_accessor,
-        //     *FIGHTER_ANIMCMD_EXPRESSION,
-        //     Hash40::new_raw(func.get_u64()),
-        //     -1
-        // );
-        WorkModule::set_int64(fighter.module_accessor, 0, *FIGHTER_STATUS_WORK_ID_INT_RESERVE_LOG_ATTACK_KIND);
-        WorkModule::on_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_ATTACK_FLAG_CHANGE_LOG);
-        fighter.change_status_jump_mini_attack(true.into());
-        return 1.into();
-    }
-    if !WorkModule::is_flag(fighter.module_accessor, *FIGHTER_STATUS_ATTACK_FLAG_RESTART) {
-        if !MotionModule::is_end(fighter.module_accessor) {
-            ryu_idkwhatthisis2(fighter);
-        }
-        else {
-            fighter.change_status(FIGHTER_STATUS_KIND_WAIT.into(), false.into());
-        }
-    }
-    else {
-        fighter.change_status(FIGHTER_STATUS_KIND_ATTACK.into(), false.into());
-    }
-    0.into()
-}
-
-#[status_script(agent = "ken", status = FIGHTER_STATUS_KIND_SPECIAL_S, condition = LUA_SCRIPT_STATUS_FUNC_INIT_STATUS)]
-unsafe fn ken_specials_init(fighter: &mut L2CFighterCommon) -> L2CValue {
-    ryu_specials_init_main(fighter)
-}
-
-#[status_script(agent = "ken", status = FIGHTER_STATUS_KIND_SPECIAL_S, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_MAIN)]
-unsafe fn ken_specials(fighter: &mut L2CFighterCommon) -> L2CValue {
-    ryu_specials_main(fighter);
-    0.into()
-}
-
-#[status_script(agent = "ken", status = FIGHTER_RYU_STATUS_KIND_SPECIAL_S_COMMAND, condition = LUA_SCRIPT_STATUS_FUNC_INIT_STATUS)]
-unsafe fn ken_specials_command_init(fighter: &mut L2CFighterCommon) -> L2CValue {
-    WorkModule::on_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_COMMON_FLAG_COMMAND);
-    ryu_specials_init_main(fighter)
-}
-
-#[status_script(agent = "ken", status = FIGHTER_RYU_STATUS_KIND_SPECIAL_S_COMMAND, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_MAIN)]
-unsafe fn ken_specials_command(fighter: &mut L2CFighterCommon) -> L2CValue {
-    ryu_specials_main(fighter);
-    0.into()
-}
-
-#[status_script(agent = "ken", status = FIGHTER_RYU_STATUS_KIND_SPECIAL_S_LOOP, condition = LUA_SCRIPT_STATUS_FUNC_INIT_STATUS)]
-unsafe fn ken_specials_loop_init(fighter: &mut L2CFighterCommon) -> L2CValue {
-    ryu_specials_loop_init_main(fighter)
-}
-
-#[status_script(agent = "ken", status = FIGHTER_RYU_STATUS_KIND_SPECIAL_S_LOOP, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_MAIN)]
-unsafe fn ken_specials_loop_main(fighter: &mut L2CFighterCommon) -> L2CValue {
-    let start_sit = WorkModule::get_int(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_S_INT_START_SITUATION);
-    if start_sit != *SITUATION_KIND_GROUND {
-        MotionModule::change_motion(
-            fighter.module_accessor,
-            Hash40::new("special_air_s"),
-            0.0,
-            1.0,
-            false,
-            0.0,
-            false,
-            false
-        );
-    }
-    else {
-        MotionModule::change_motion(
-            fighter.module_accessor,
-            Hash40::new("special_s"),
-            0.0,
-            1.0,
-            false,
-            0.0,
-            false,
-            false
-        );
-    }
-    let eff = if !MotionModule::is_flip(fighter.module_accessor) {
-        Hash40::new("ken_tatsumaki_wind_r")
-    }
-    else {
-        Hash40::new("ken_tatsumaki_wind_l")
-    };
-    fighter.clear_lua_stack();
-    lua_args!(fighter, MA_MSC_EFFECT_REQUEST_FOLLOW, eff, hash40("rot"), 0.0, 1.5, 0.0, 0.0, 0.0, 0.0, 1.0, false, *EFFECT_SUB_ATTRIBUTE_SYNC_STOP, 0, -1);
-    sv_module_access::effect(fighter.lua_state_agent);
-    let spineffect = fighter.pop_lua_stack(1).get_u32();
-    WorkModule::set_int(fighter.module_accessor, spineffect as i32, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_S_INT_EFFECT_HANDLE);
-    if !WorkModule::is_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_COMMON_FLAG_COMMAND) {
-        let alpha = WorkModule::get_param_float(fighter.module_accessor, hash40("param_special_s"), hash40("wind_alpha")) * 0.01;
-        EffectModule::set_alpha(fighter.module_accessor, spineffect, alpha);
-    }
-    else {
-        let alpha = WorkModule::get_param_float(fighter.module_accessor, hash40("param_special_s"), hash40("command_wind_alpha")) * 0.01;
-        EffectModule::set_alpha(fighter.module_accessor, spineffect, alpha);
-    }
-    fighter.sub_shift_status_main(L2CValue::Ptr(ken_specials_loop_main_loop as *const () as _))
-}
-
-unsafe extern "C" fn ken_specials_loop_main_loop(fighter: &mut L2CFighterCommon) -> L2CValue {
-    if fighter.sub_transition_group_check_air_cliff().get_bool() {
-        return 1.into();
-    }
-    if !StatusModule::is_changing(fighter.module_accessor) {
-        if StatusModule::is_situation_changed(fighter.module_accessor) {
-            if fighter.global_table[SITUATION_KIND].get_i32() != *SITUATION_KIND_GROUND {
-                GroundModule::correct(fighter.module_accessor, GroundCorrectKind(*GROUND_CORRECT_KIND_AIR));
-            }
-            else {
-                if WorkModule::is_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_COMMON_FLAG_COMMAND) {
-                    WorkModule::set_float(fighter.module_accessor, 10.0, *FIGHTER_INSTANCE_WORK_ID_FLOAT_LANDING_FRAME);
-                    fighter.change_status(FIGHTER_STATUS_KIND_LANDING_FALL_SPECIAL.into(), false.into());
-                    return 1.into();
-                }
-                GroundModule::correct(fighter.module_accessor, GroundCorrectKind(*GROUND_CORRECT_KIND_GROUND_CLIFF_STOP));
-            }
-        }
-    }
-    else {
-        if fighter.global_table[SITUATION_KIND].get_i32() != *SITUATION_KIND_GROUND {
-            GroundModule::correct(fighter.module_accessor, GroundCorrectKind(*GROUND_CORRECT_KIND_AIR));
-        }
-        else {
-            GroundModule::correct(fighter.module_accessor, GroundCorrectKind(*GROUND_CORRECT_KIND_GROUND_CLIFF_STOP));
-            let start_sit = WorkModule::get_int(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_S_INT_START_SITUATION);
-            if start_sit != *SITUATION_KIND_GROUND {
-                KineticModule::change_kinetic(fighter.module_accessor, *FIGHTER_KINETIC_TYPE_GROUND_STOP);
-                sv_kinetic_energy!(
-                    set_accel,
-                    fighter,
-                    FIGHTER_KINETIC_ENERGY_ID_STOP,
-                    0.0,
-                    0.0
-                );
-                sv_kinetic_energy!(
-                    set_brake,
-                    fighter,
-                    FIGHTER_KINETIC_ENERGY_ID_STOP,
-                    0.0,
-                    0.0
-                );
-                sv_kinetic_energy!(
-                    set_stable_speed,
-                    fighter,
-                    FIGHTER_KINETIC_ENERGY_ID_STOP,
-                    0.0,
-                    0.0
-                );
-            }
-        }
-    }
-    if MotionModule::is_end(fighter.module_accessor) {
-        WorkModule::dec_int(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_S_INT_LOOP_COUNT);
-        let loop_count = WorkModule::get_int(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_S_INT_LOOP_COUNT);
-        if loop_count > 0 {
-            let start_sit = WorkModule::get_int(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_S_INT_START_SITUATION);
-            if start_sit != *SITUATION_KIND_GROUND {
-                MotionModule::change_motion(
-                    fighter.module_accessor,
-                    Hash40::new("special_air_s"),
-                    0.0,
-                    1.0,
-                    false,
-                    0.0,
-                    false,
-                    false
-                );
-            }
-            else {
-                MotionModule::change_motion(
-                    fighter.module_accessor,
-                    Hash40::new("special_s"),
-                    0.0,
-                    1.0,
-                    false,
-                    0.0,
-                    false,
-                    false
-                );
-            }
-        }
-        else {
-            fighter.change_status(FIGHTER_RYU_STATUS_KIND_SPECIAL_S_END.into(), false.into());
-        }
-    }
-    if fighter.global_table[SITUATION_KIND].get_i32() != *SITUATION_KIND_GROUND {
-        WorkModule::off_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_S_FLAG_GROUND);
-    }
-    else {
-        WorkModule::on_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_S_FLAG_GROUND);
-    }
-    0.into()
-}
-
-#[status_script(agent = "ken", status = FIGHTER_RYU_STATUS_KIND_SPECIAL_HI_COMMAND, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_MAIN)]
-unsafe fn ken_specialhi_command(fighter: &mut L2CFighterCommon) -> L2CValue {
-    ken_specialhi_main(fighter);
-    0.into()
-}
-
-unsafe extern "C" fn ken_specialhi_main(fighter: &mut L2CFighterCommon) -> L2CValue {
-    let command = WorkModule::is_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_COMMON_FLAG_COMMAND);
-    let mot = if fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_GROUND {
-        if command {
-            Hash40::new("special_hi_command")
-        }
-        else {
-            Hash40::new("special_hi")
-        }
-    }
-    else {
-        if command {
-            Hash40::new("special_air_hi_command")
-        }
-        else {
-            Hash40::new("special_air_hi")
-        }
-    };
-    MotionModule::change_motion(
-        fighter.module_accessor,
-        mot,
-        0.0,
-        1.0,
-        false,
-        0.0,
-        false,
-        false
-    );
-    ItemModule::set_change_status_event(fighter.module_accessor, false);
-    if !StopModule::is_stop(fighter.module_accessor) {
-        ken_specialhi_substatus(fighter, false.into());
-    }
-    fighter.global_table[SUB_STATUS].assign(&L2CValue::Ptr(ken_specialhi_substatus as *const () as _));
-    fighter.sub_shift_status_main(L2CValue::Ptr(ken_specialhi_main_loop as *const () as _))
-}
-
-unsafe extern "C" fn ken_specialhi_substatus(fighter: &mut L2CFighterCommon, param_1: L2CValue) -> L2CValue {
-    if !param_1.get_bool() {
-        if WorkModule::is_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_HI_FLAG_REVERSE_LR) {
-            WorkModule::off_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_HI_FLAG_REVERSE_LR);
-            let stick_x = fighter.global_table[STICK_X].get_f32();
-            let lr_stick_x = WorkModule::get_param_float(fighter.module_accessor, hash40("param_special_hi"), hash40("lr_stick_x"));
-            if lr_stick_x < stick_x.abs() {
-                PostureModule::set_stick_lr(fighter.module_accessor, 0.0);
-                PostureModule::update_rot_y_lr(fighter.module_accessor);
-            }
-        }
-    }
-    else {
-        WorkModule::inc_int(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_COMMON_INT_BUTTON_ON_TIMER);
-    }
-    0.into()
-}
-
-unsafe extern "C" fn ken_specialhi_main_loop(fighter: &mut L2CFighterCommon) -> L2CValue {
-    if VarModule::is_flag(fighter.battle_object, ken::status::flag::SPECIAL_HI_CHANGE_REPPA)
-    && VarModule::is_flag(fighter.battle_object, ken::instance::flag::V_TRIGGER)
-    && WorkModule::get_int(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_COMMON_INT_STRENGTH) == *FIGHTER_RYU_STRENGTH_S
-    && fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_GROUND {
-        MotionModule::change_motion_inherit_frame(
-            fighter.module_accessor,
-            Hash40::new("special_hi_reppa"),
-            -1.0,
-            1.0,
-            0.0,
-            false,
-            false
-        );
-        GroundModule::set_correct(fighter.module_accessor, GroundCorrectKind(*GROUND_CORRECT_KIND_GROUND_CLIFF_STOP_ATTACK));
-        VarModule::off_flag(fighter.battle_object, ken::status::flag::SPECIAL_HI_CHANGE_REPPA);
-    }
-    if StatusModule::is_changing(fighter.module_accessor)
-    || StatusModule::is_situation_changed(fighter.module_accessor) {
-        if fighter.global_table[SITUATION_KIND].get_i32() != *SITUATION_KIND_GROUND {
-            // if MotionModule::motion_kind(fighter.module_accessor) == hash40("special_hi_reppa") {
-            //     MotionModule::change_motion_inherit_frame(
-            //         fighter.module_accessor,
-            //         Hash40::new("special_hi_command"),
-            //         -1.0,
-            //         1.0,
-            //         0.0,
-            //         false,
-            //         false
-            //     );
-            // }
-            GroundModule::correct(fighter.module_accessor, GroundCorrectKind(*GROUND_CORRECT_KIND_AIR));
-            if !StatusModule::is_changing(fighter.module_accessor) {
-                if VarModule::is_flag(fighter.battle_object, ken::status::flag::SPECIAL_HI_CHANGE_REPPA) {
-                    KineticModule::change_kinetic(fighter.module_accessor, *FIGHTER_KINETIC_TYPE_MOTION_AIR);
-                }
-                else {
-                    KineticModule::change_kinetic(fighter.module_accessor, *FIGHTER_KINETIC_TYPE_AIR_STOP);
-                    sv_kinetic_energy!(
-                        reset_energy,
-                        fighter,
-                        FIGHTER_KINETIC_ENERGY_ID_MOTION,
-                        ENERGY_MOTION_RESET_TYPE_AIR_TRANS,
-                        0.0,
-                        0.0,
-                        0.0,
-                        0.0,
-                        0.0
-                    );
-                    KineticModule::enable_energy(fighter.module_accessor, *FIGHTER_KINETIC_ENERGY_ID_MOTION);
-                    let start_accel_y = WorkModule::get_param_int(fighter.module_accessor, hash40("param_special_hi"), hash40("start_accel_y"));
-                    sv_kinetic_energy!(
-                        set_accel,
-                        fighter,
-                        FIGHTER_KINETIC_ENERGY_ID_GRAVITY,
-                        -start_accel_y
-                    );
-                }
-            }
-        }
-        else {
-            GroundModule::correct(fighter.module_accessor, GroundCorrectKind(*GROUND_CORRECT_KIND_GROUND));
-            if !StatusModule::is_changing(fighter.module_accessor) {
-                KineticModule::change_kinetic(fighter.module_accessor, *FIGHTER_KINETIC_TYPE_GROUND_STOP);
-                sv_kinetic_energy!(
-                    reset_energy,
-                    fighter,
-                    FIGHTER_KINETIC_ENERGY_ID_MOTION,
-                    ENERGY_MOTION_RESET_TYPE_GROUND_TRANS,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0,
-                    0.0
-                );
-                KineticModule::enable_energy(fighter.module_accessor, *FIGHTER_KINETIC_ENERGY_ID_MOTION);              
-            }
-        }
-    }
-    if WorkModule::is_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_HI_FLAG_JUMP) {
-        WorkModule::off_flag(fighter.module_accessor, *FIGHTER_RYU_STATUS_WORK_ID_SPECIAL_HI_FLAG_JUMP);
-        fighter.change_status(FIGHTER_RYU_STATUS_KIND_SPECIAL_HI_JUMP.into(), false.into());
-    }
-    0.into()
-}
-
-#[status_script(agent = "ken", status = FIGHTER_STATUS_KIND_SPECIAL_LW, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_PRE)]
-unsafe fn ken_speciallw_pre(fighter: &mut L2CFighterCommon) -> L2CValue {
-    StatusModule::init_settings(
-        fighter.module_accessor,
-        SituationKind(*SITUATION_KIND_NONE),
-        *FIGHTER_KINETIC_TYPE_UNIQ,
-        *GROUND_CORRECT_KIND_KEEP as u32,
-        GroundCliffCheckKind(*GROUND_CLIFF_CHECK_KIND_NONE),
-        true,
-        *FIGHTER_STATUS_WORK_KEEP_FLAG_NONE_FLAG,
-        *FIGHTER_STATUS_WORK_KEEP_FLAG_NONE_INT,
-        *FIGHTER_STATUS_WORK_KEEP_FLAG_NONE_FLOAT,
-        0
-    );
-    FighterStatusModuleImpl::set_fighter_status_data(
-        fighter.module_accessor,
-        false,
-        *FIGHTER_TREADED_KIND_NO_REAC,
-        false,
-        false,
-        false,
-        (*FIGHTER_LOG_MASK_FLAG_ACTION_TRIGGER_ON | *FIGHTER_LOG_MASK_FLAG_ATTACK_KIND_SPECIAL_LW) as u64,
-        0,
-        *FIGHTER_POWER_UP_ATTACK_BIT_SPECIAL_LW as u32,
-        0
-    );
-    0.into()
-}
-
-#[status_script(agent = "ken", status = FIGHTER_STATUS_KIND_SPECIAL_LW, condition = LUA_SCRIPT_STATUS_FUNC_INIT_STATUS)]
-unsafe fn ken_speciallw_init(fighter: &mut L2CFighterCommon) -> L2CValue {
-    sv_kinetic_energy!(
-        reset_energy,
-        fighter,
-        FIGHTER_KINETIC_ENERGY_ID_STOP,
-        ENERGY_STOP_RESET_TYPE_AIR,
-        0.0,
-        0.0,
-        0.0,
-        0.0
-    );
-    KineticModule::enable_energy(fighter.module_accessor, *FIGHTER_KINETIC_ENERGY_ID_STOP);
-    let reset_type = if fighter.global_table[SITUATION_KIND].get_i32() != *SITUATION_KIND_GROUND {
-        ENERGY_MOTION_RESET_TYPE_AIR_TRANS
-    }
-    else {
-        ENERGY_MOTION_RESET_TYPE_GROUND_TRANS
-    };
-    sv_kinetic_energy!(
-        reset_energy,
-        fighter,
-        FIGHTER_KINETIC_ENERGY_ID_MOTION,
-        reset_type,
-        0.0,
-        0.0,
-        0.0,
-        0.0,
-        0.0
-    );
-    KineticModule::enable_energy(fighter.module_accessor, *FIGHTER_KINETIC_ENERGY_ID_MOTION);
-    sv_kinetic_energy!(
-        reset_energy,
-        fighter,
-        FIGHTER_KINETIC_ENERGY_ID_GRAVITY,
-        ENERGY_GRAVITY_RESET_TYPE_GRAVITY,
-        0.0,
-        0.0,
-        0.0,
-        0.0
-    );
-    if fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_GROUND {
-        KineticModule::enable_energy(fighter.module_accessor, *FIGHTER_KINETIC_ENERGY_ID_GRAVITY);
-    }
-    else {
-        let accel_y = WorkModule::get_param_float(fighter.module_accessor, hash40("param_special_lw"), hash40("accel_y"));
-        sv_kinetic_energy!(
-            set_accel,
-            fighter,
-            FIGHTER_KINETIC_ENERGY_ID_GRAVITY,
-            -accel_y
-        );
-        let max_y = WorkModule::get_param_float(fighter.module_accessor, hash40("param_special_lw"), hash40("max_speed_y"));
-        sv_kinetic_energy!(
-            set_limit_speed,
-            fighter,
-            FIGHTER_KINETIC_ENERGY_ID_GRAVITY,
-            max_y
-        );
-        KineticModule::enable_energy(fighter.module_accessor, *FIGHTER_KINETIC_ENERGY_ID_GRAVITY);
-    }
-    0.into()
-}
-
-#[status_script(agent = "ken", status = FIGHTER_STATUS_KIND_SPECIAL_LW, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_MAIN)]
-unsafe fn ken_speciallw_main(fighter: &mut L2CFighterCommon) -> L2CValue {
-    ControlModule::clear_command_one(fighter.module_accessor, *FIGHTER_PAD_COMMAND_CATEGORY1, *FIGHTER_PAD_CMD_CAT1_SPECIAL_LW);
-    if fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_GROUND {
-        GroundModule::set_correct(fighter.module_accessor, GroundCorrectKind(*GROUND_CORRECT_KIND_GROUND_CLIFF_STOP));
-        let mot = if VarModule::get_float(fighter.battle_object, ken::instance::float::V_GAUGE) < 900.0
-        || VarModule::is_flag(fighter.battle_object, ken::instance::flag::V_TRIGGER) {
-            VarModule::set_int(
-                fighter.battle_object,
-                ken::instance::int::SPECIAL_LW_TYPE,
-                ken::SPECIAL_LW_TYPE_QUICK_STEP
-            );
-            VarModule::set_int(
-                fighter.battle_object,
-                ken::instance::int::QUICK_STEP_STATE,
-                ken::QUICK_STEP_STATE_RUN
-            );
-            Hash40::new("run")
-        }
-        else {
-            VarModule::set_int(
-                fighter.battle_object,
-                ken::instance::int::SPECIAL_LW_TYPE,
-                ken::SPECIAL_LW_TYPE_HEAT_RUSH
-            );
-            Hash40::new("special_lw_step_f")
-        };
-        MotionModule::change_motion(
-            fighter.module_accessor,
-            mot,
-            0.0,
-            1.0,
-            false,
-            0.0,
-            false,
-            false
-        );
-    }
-    else {
-        if VarModule::get_float(fighter.battle_object, ken::instance::float::V_GAUGE) < 900.0
-        || VarModule::is_flag(fighter.battle_object, ken::instance::flag::V_TRIGGER) {
-            KineticModule::change_kinetic(fighter.module_accessor, *FIGHTER_KINETIC_TYPE_UNIQ);
-            macros::SET_SPEED_EX(fighter, 0.8, 0.2, *KINETIC_ENERGY_RESERVE_ATTRIBUTE_MAIN);
-            VarModule::set_int(
-                fighter.battle_object,
-                ken::instance::int::SPECIAL_LW_TYPE,
-                ken::SPECIAL_LW_TYPE_QUICK_STEP
-            );
-            VarModule::set_int(
-                fighter.battle_object,
-                ken::instance::int::QUICK_STEP_STATE,
-                ken::QUICK_STEP_STATE_DISABLE
-            );
-        }
-        else {
-            VarModule::set_int(
-                fighter.battle_object,
-                ken::instance::int::SPECIAL_LW_TYPE,
-                ken::SPECIAL_LW_TYPE_HEAT_RUSH
-            );
-        }
-        MotionModule::change_motion(
-            fighter.module_accessor,
-            Hash40::new("special_air_lw_step_f"),
-            0.0,
-            1.0,
-            false,
-            0.0,
-            false,
-            false
-        );
-    }
-    if VarModule::get_int(fighter.battle_object, ken::instance::int::SPECIAL_LW_TYPE) == ken::SPECIAL_LW_TYPE_QUICK_STEP {
-        fighter.sub_shift_status_main(L2CValue::Ptr(ken_quickstep_loop as *const () as _))
-    }
-    else {
-        KineticModule::change_kinetic(fighter.module_accessor, *FIGHTER_KINETIC_TYPE_RESET);
-        HitModule::set_status_all(fighter.module_accessor, HitStatus(*HIT_STATUS_XLU), 0);
-        SlowModule::set_whole(fighter.module_accessor, 6, 0);
-        macros::SLOW_OPPONENT(fighter, 100.0, 12.0);
-        macros::FILL_SCREEN_MODEL_COLOR(fighter, 0, 3, 0.2, 0.2, 0.2, 0, 0, 0, 1, 1, *smash::lib::lua_const::EffectScreenLayer::GROUND, 205);
-        let target_id = VarModule::get_int(fighter.battle_object, commons::instance::int::TARGET_ID) as u32;
-        if sv_battle_object::is_active(target_id) {
-            let target_boma = sv_battle_object::module_accessor(target_id);
-            let mut diff_x = PostureModule::pos_x(target_boma) - PostureModule::pos_x(fighter.module_accessor);
-            if (diff_x > 0.0 && PostureModule::lr(fighter.module_accessor) < 0.0)
-            || (diff_x < 0.0 && PostureModule::lr(fighter.module_accessor) > 0.0) {
-                PostureModule::reverse_lr(fighter.module_accessor);
-                PostureModule::reverse_rot_y_lr(fighter.module_accessor);
-            }
-            diff_x = diff_x.abs();
-            if diff_x > 50.0 {
-                diff_x -= 5.0;
-            }
-            else {
-                diff_x = 0.0;
-            }
-            VarModule::set_float(fighter.battle_object, ken::instance::float::DIFF_X, diff_x);
-            VarModule::set_int(fighter.battle_object, commons::instance::int::TARGET_ID, 0);
-        }
-        else {
-            VarModule::set_float(fighter.battle_object, ken::instance::float::DIFF_X, 0.0);
-        }
-        macros::PLAY_SE(fighter, Hash40::new("se_ken_special_l01"));
-        macros::PLAY_SE(fighter, Hash40::new("vc_ken_special_l01"));
-        VarModule::on_flag(fighter.battle_object, ken::instance::flag::V_TRIGGER);
-        VarModule::set_float(fighter.battle_object, ken::instance::float::V_GAUGE, 0.0);
-        fighter.sub_shift_status_main(L2CValue::Ptr(ken_heatrush_loop as *const () as _))
-    }
-}
-
-unsafe extern "C" fn ken_quickstep_loop(fighter: &mut L2CFighterCommon) -> L2CValue {
-    if MotionModule::motion_kind(fighter.module_accessor) == hash40("run") {
-        if CancelModule::is_enable_cancel(fighter.module_accessor) {
-            if fighter.sub_transition_group_check_ground_jump_mini_attack().get_bool()
-            || fighter.sub_transition_group_check_ground_item().get_bool()
-            || fighter.sub_transition_group_check_ground_catch().get_bool()
-            || fighter.sub_transition_group_check_ground_escape().get_bool()
-            || fighter.sub_transition_group_check_ground_guard().get_bool()
-            || fighter.sub_transition_group_check_ground_attack().get_bool()
-            || fighter.sub_transition_group_check_ground_jump().get_bool() {
-                return 0.into();
-            }
-        }
-    }
-    else if CancelModule::is_enable_cancel(fighter.module_accessor) {
-        if fighter.sub_wait_ground_check_common(false.into()).get_bool()
-        || fighter.sub_air_check_fall_common().get_bool() {
-            return 0.into();
-        }
-    }
-    if VarModule::is_flag(fighter.battle_object, ken::status::flag::SPECIAL_LW_STEP_KICK) {
-        MotionModule::change_motion(
-            fighter.module_accessor,
-            Hash40::new("attack_s3_s_w"),
-            0.0,
-            1.0,
-            false,
-            0.0,
-            false,
-            false
-        );
-        VarModule::off_flag(fighter.battle_object, ken::status::flag::SPECIAL_LW_STEP_KICK);
-    }
-    if MotionModule::motion_kind(fighter.module_accessor) == hash40("special_air_lw_step_f") {
-        if fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_GROUND {
-            fighter.change_status(FIGHTER_STATUS_KIND_LANDING.into(), false.into());
-        }
-    }
-    if MotionModule::is_end(fighter.module_accessor) {
-        if fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_AIR {
-            fighter.change_status(FIGHTER_STATUS_KIND_FALL.into(), false.into());
-        }
-        else if fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_GROUND {
-            fighter.change_status(FIGHTER_STATUS_KIND_WAIT.into(), false.into());
-        }
-    }
-    0.into()
-}
-
-unsafe extern "C" fn ken_heatrush_loop(fighter: &mut L2CFighterCommon) -> L2CValue {
-    if CancelModule::is_enable_cancel(fighter.module_accessor) {
-        fighter.sub_wait_ground_check_common(false.into());
-        fighter.sub_air_check_fall_common();
-    }
-    if MotionModule::is_end(fighter.module_accessor) {
-        if fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_AIR {
-            fighter.change_status(FIGHTER_STATUS_KIND_FALL.into(), false.into());
-        }
-        else if fighter.global_table[SITUATION_KIND].get_i32() == *SITUATION_KIND_GROUND {
-            fighter.change_status(FIGHTER_STATUS_KIND_WAIT.into(), false.into());
-        }
-    }
-    0.into()
-}
+mod dash_back;
+mod attack;
+mod special_s;
+mod special_s_loop;
+mod special_hi;
+mod special_lw;
 
 pub fn install() {
-    install_status_scripts!(
-        ken_dashback_pre,
-        ken_dashback_main,
-        ken_attack_main,
-        ken_specials_init,
-        ken_specials,
-        ken_specials_command_init,
-        ken_specials_command,
-        ken_specials_loop_init,
-        ken_specials_loop_main,
-        ken_specialhi_command,
-        ken_speciallw_pre,
-        ken_speciallw_init,
-        ken_speciallw_main
-    );
+    dash_back::install();
+    attack::install();
+    special_s::install();
+    special_s_loop::install();
+    special_hi::install();
+    special_lw::install();
 }
