+++ src/fighter/ken/status/dash_back.rs
@@ -1,0 +1,25 @@
+use {
+    smash::{
+        lua2cpp::L2CFighterCommon,
+        lib::{lua_const::*, L2CValue}
+    },
+    smashline::*,
+    crate::fighter::common::common_status::dash::*
+};
+
+#[status_script(agent = "ken", status = FIGHTER_RYU_STATUS_KIND_DASH_BACK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_PRE)]
+unsafe fn ken_dashback_pre(fighter: &mut L2CFighterCommon) -> L2CValue {
+    fgc_dashback_pre(fighter)
+}
+
+#[status_script(agent = "ken", status = FIGHTER_RYU_STATUS_KIND_DASH_BACK, condition = LUA_SCRIPT_STATUS_FUNC_STATUS_MAIN)]
+unsafe fn ken_dashback_main(fighter: &mut L2CFighterCommon) -> L2CValue {
+    fgc_dashback_main(fighter)
+}
+
+pub fn install() {
+    install_status_scripts!(
+        ken_dashback_pre,
+        ken_dashback_main
+    );
+}
